+++ R/pipe.R
@@ -19,7 +19,7 @@
 #' Deferred Evaluation Pipe
 #'
 #' Pipe a possibly unresolved value forward into a function. The piped expression
-#'     should be wrapped in \code{resolve()}.
+#'     should be wrapped in \code{.()}.
 #'
 #' @param x a 'mirai' or mirai value at \code{$data} that is possibly an
 #'     'unresolvedValue'.
@@ -41,14 +41,13 @@
 #'     Supports stringing together a series of piped expressions (as per
 #'     the below example).
 #'
-#'     Wrapping a piped expression in \code{\link{resolve}} does not force
-#'     immediate resolution of the expression, but ensures that the return value
-#'     is always an 'unresolvedExpr' or 'resolvedExpr' as the case may be,
+#'     Wrap a piped expression in \code{.()} to ensure that the return value is
+#'     always an 'unresolvedExpr' or 'resolvedExpr' as the case may be,
 #'     otherwise if 'x' is already resolved, the evaluated result would be
 #'     returned directly.
 #'
-#'     \code{\link{unresolved}} may be used on the expression returned by
-#'     \code{\link{resolve}} or its \code{$data} element to test for resolution.
+#'     \code{\link{unresolved}} may be used on an expression or its \code{$data}
+#'     element to test for resolution.
 #'
 #' @section Usage:
 #'
@@ -68,7 +67,7 @@
 #' # Only run examples in interactive R sessions
 #'
 #' m <- mirai({Sys.sleep(0.5); 1})
-#' b <- resolve(m %>>% c(2, 3) %>>% as.character)
+#' b <- .(m %>>% c(2, 3) %>>% as.character)
 #' unresolved(b)
 #' b
 #' b$data
@@ -80,6 +79,7 @@
 #'
 #' }
 #'
+#' @rdname deferred-execution-pipe
 #' @export
 #'
 `%>>%` <- function(x, f) {
@@ -108,10 +108,10 @@
   }
 }
 
-#' @rdname grapes-greater-than-greater-than-grapes
+#' @rdname deferred-execution-pipe
 #' @export
 #'
-resolve <- function(expr)
+. <- function(expr)
   if (inherits(expr, c("unresolvedExpr", "resolvedExpr"))) expr else
     `class<-`(`[[<-`(new.env(hash = FALSE), "data", expr), "resolvedExpr")